Updated the environment to run ansible on docker

+++ docker/entrypoint.sh
@@ -14,6 +14,24 @@ main() {
         return 1
     }
 
+    create_ansible_environment || {
+        log_err "Failed to create Ansible environment"
+        return 1
+    }
+
+    prepare_ssh_key || {
+        log_err "Failed to prepare ssh-keys to authenticate"
+        return 1
+    }
+
+    ${ANSIBLE_DIRECTORY}/${PYTHON_VIRTUALENV_DIRECTORY}/bin/ansible-playbook -l target-host -i production site.yml
+}
+
+create_ansible_environment() {
+
+    local prepared_flag_file="${ANSIBLE_DIRECTORY}/${PYTHON_VIRTUALENV_DIRECTORY}/.ansible_environment_has_already_prepared"
+    [[ -f "$prepared_flag_file" ]] && return
+
     python3 -m "$PYTHON_VIRTUALENV_DIRECTORY" "${PYTHON_VIRTUALENV_DIRECTORY}/" || {
         log_err "Failed to install python virtual env in $ANSIBLE_DIRECTORY"
         return 1
@@ -32,14 +50,7 @@ main() {
     #/opt/ansible/venv/bin/ansible-galaxy 
     ${ANSIBLE_DIRECTORY}/${PYTHON_VIRTUALENV_DIRECTORY}/bin/ansible-galaxy install -r requirements.yml
 
-    prepare_ssh_key || {
-        log_err "Failed to prepare ssh-keys to authenticate"
-        return 1
-    }
-
-    ${ANSIBLE_DIRECTORY}/${PYTHON_VIRTUALENV_DIRECTORY}/bin/ansible-playbook -l target-host -i production site.yml
-
-    tail -f /dev/null
+    touch "$prepared_flag_file"
 }
 
 prepare_ssh_key() {
